MTL files and folder created
