--- __init__.py
@@ -1,4 +1,0 @@
-from stlLexer import stlLexer
-from stlParser import stlParser
-from stl import STLAbstractSyntaxTreeExtractor
-from stl import Operation, RelOperation, STLFormula, Trace
