+++ mtl/__init__.py
@@ -1,0 +1,4 @@
+from mtlLexer import mtlLexer
+from mtlParser import mtlParser
+from mtl import MTLAbstractSyntaxTreeExtractor
+from mtl import Operation, RelOperation, MTLFormula, Trace
